requirements.txt
requests
python-dotenv
playwright
moviepy==1.0.3
Pillow<10.0.0
gTTS
